Check if the auto-compound task is already scheduled
