+++ demo/package.json
@@ -5,6 +5,9 @@
   "main": "index.js",
   "author": "",
   "license": "ISC",
+  "scripts": {
+    "compound-task": "babel-node src/scheduleDynamicDispatchTaskToAutoCompound.js"
+  },
   "dependencies": {
     "@oak-network/api-augment": "file:../packages/api-augment/build",
     "@oak-network/types": "file:../packages/types/build",
@@ -12,5 +15,10 @@
     "@polkadot/wasm-crypto": "^7.3.1",
     "bn.js": "^5.2.1",
     "lodash": "^4.17.21"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.20.12",
+    "@babel/node": "^7.20.2",
+    "@babel/preset-env": "^7.20.2"
   }
 }
