+++ demo/src/scheduleDynamicDispatchTaskToAutoCompound.js
@@ -1,164 +1,21 @@
-const {
-  find,
-  isNil,
-  isEmpty,
-  keys,
-  each,
-  isUndefined,
-  findIndex,
-} = require("lodash");
-const BN = require("bn.js");
-require("@oak-network/api-augment");
-const { rpc, types } = require("@oak-network/types");
-const { ApiPromise, WsProvider, Keyring } = require("@polkadot/api");
-const { waitReady } = require("@polkadot/wasm-crypto");
+import _ from "lodash";
+import { ApiPromise, WsProvider } from "@polkadot/api";
+import BN from "bn.js";
+import "@oak-network/api-augment";
+import { rpc, types } from "@oak-network/types";
+
+import {
+  sendExtrinsic,
+  listenEvents,
+  getTaskIdInTaskScheduledEvent,
+  findEvent,
+  getNextTenMinutesTimestamp,
+  getKeyringPair,
+} from "./utils";
 
 const MIN_ACCOUNT_BALANCE = new BN(100);
 const AUTO_COMPOUND_AMOUNT = new BN(10000000000);
 const AUTO_COMPOUND_INTERVAL = 600;
-
-function getNextTenMinutesTimestamp() {
-  const currentTime = new Date();
-  const currentMinutes = currentTime.getMinutes();
-  const remainingMinutes = 10 - (currentMinutes % 10);
-
-  const nextTenMinutes = new Date(
-    currentTime.getTime() + remainingMinutes * 60000,
-  );
-
-  nextTenMinutes.setSeconds(0);
-  nextTenMinutes.setMilliseconds(0);
-
-  return nextTenMinutes.getTime();
-}
-
-const getKeyringPair = async (ss58Format) => {
-  await waitReady();
-  if (isEmpty(process.env.SENDER_MNEMONIC)) {
-    throw new Error("The SENDER_MNEMONIC environment variable is not set.");
-  }
-  // Generate sender keyring pair from mnemonic
-  const keyring = new Keyring({ ss58Format, type: "sr25519" });
-  const keyringPair = keyring.addFromMnemonic(process.env.SENDER_MNEMONIC);
-  return keyringPair;
-};
-
-const sendExtrinsic = (extrinsic, api, keyringPair) =>
-  new Promise((resolve, reject) => {
-    const signAndSend = async () => {
-      const unsub = await extrinsic.signAndSend(keyringPair, (result) => {
-        const { status, events, dispatchError } = result;
-        console.log("status.type: ", status.type);
-
-        if (status?.isFinalized) {
-          unsub();
-          if (!isNil(dispatchError)) {
-            reject(dispatchError);
-          }
-
-          const event = find(events, ({ event: eventData }) =>
-            api.events.system.ExtrinsicSuccess.is(eventData),
-          );
-          if (event) {
-            resolve({
-              blockHash: status?.asFinalized?.toString(),
-              events,
-              extrinsicHash: extrinsic.hash,
-            });
-          } else {
-            reject(new Error("The event.ExtrinsicSuccess is not found"));
-          }
-        }
-      });
-    };
-    signAndSend();
-  });
-
-const findEvent = (events, section, method) =>
-  events.find((e) => e.event.section === section && e.event.method === method);
-
-const getTaskIdInTaskScheduledEvent = (event) =>
-  Buffer.from(event.event.data.taskId).toString();
-
-const listenEvents = async (
-  api,
-  section,
-  method,
-  conditions,
-  timeout = undefined,
-) =>
-  new Promise((resolve) => {
-    let unsub = null;
-    let timeoutId = null;
-
-    if (timeout) {
-      timeoutId = setTimeout(() => {
-        unsub();
-        resolve(null);
-      }, timeout);
-    }
-
-    const listenSystemEvents = async () => {
-      unsub = await api.query.system.events((events) => {
-        const foundEventIndex = findIndex(events, ({ event }) => {
-          const { section: eventSection, method: eventMethod, data } = event;
-          if (eventSection !== section || eventMethod !== method) {
-            return false;
-          }
-
-          if (!isUndefined(conditions)) {
-            return true;
-          }
-
-          let conditionPassed = true;
-          each(keys(conditions), (key) => {
-            if (conditions[key] === data[key]) {
-              conditionPassed = false;
-            }
-          });
-
-          return conditionPassed;
-        });
-
-        if (foundEventIndex !== -1) {
-          const foundEvent = events[foundEventIndex];
-          const {
-            event: {
-              section: eventSection,
-              method: eventMethod,
-              typeDef,
-              data: eventData,
-            },
-            phase,
-          } = foundEvent;
-
-          // Print out the name of the event found
-          console.log(
-            `\t${eventSection}:${eventMethod}:: (phase=${phase.toString()})`,
-          );
-
-          // Loop through the conent of the event, displaying the type and data
-          eventData.forEach((data, index) => {
-            console.log(`\t\t\t${typeDef[index].type}: ${data.toString()}`);
-          });
-
-          unsub();
-
-          if (timeoutId) {
-            clearTimeout(timeoutId);
-          }
-
-          resolve({
-            events,
-            foundEvent,
-            foundEventIndex,
-          });
-        }
-      });
-    };
-
-    listenSystemEvents().catch(console.log);
-  });
 
 async function main() {
   const providerUrl =
@@ -202,18 +59,18 @@ async function main() {
   const delegatorState = await api.query.parachainStaking.delegatorState(
     delegatorWalletAddress,
   );
-  let foundDelegation = null;
+  let foundDelegation;
   console.log("delegatorState: ", delegatorState.toHuman());
   console.log("delegatorWalletAddress: ", delegatorWalletAddress);
   if (delegatorState.isSome) {
     const { delegations } = delegatorState.unwrap();
-    foundDelegation = find(delegations, ({ owner }) => {
+    foundDelegation = _.find(delegations, ({ owner }) => {
       console.log("owner: ", owner.toHex());
       return owner.toHex() === collatorWalletAddress;
     });
   }
 
-  if (!foundDelegation) {
+  if (_.isUndefined(foundDelegation)) {
     // Get Delegation params
     const minDelegationStake = api.consts.parachainStaking.minDelegation;
     const candidateInfo = await api.query.parachainStaking.candidateInfo(
@@ -268,7 +125,33 @@ async function main() {
   }
 
   console.log("\n4. Scheduling a dynamic task to auto compound...");
-  // TODO: How to check if the task is already scheduled?
+
+  // Check if the task is already scheduled
+  const entries = await api.query.automationTime.accountTasks.entries(
+    delegatorWalletAddress,
+  );
+
+  const entry = _.find(entries, ([, value]) => {
+    const task = value.unwrap();
+    const { action } = task;
+    if (!action.isDynamicDispatch) {
+      return false;
+    }
+    const { encodedCall } = action.asDynamicDispatch;
+    const extrinsicCall = api.createType("Call", encodedCall);
+    const { method, section, args } = extrinsicCall;
+    if (section !== "parachainStaking" || method !== "delegatorBondMore") {
+      return false;
+    }
+    const [candidate] = args;
+    return collatorWalletAddress === candidate.toHex();
+  });
+
+  if (!_.isUndefined(entry)) {
+    console.log("The task is already scheduled.");
+    return;
+  }
+
   // Create bondMoreExtrinsic
   const bondMoreExtrinsic = api.tx.parachainStaking.delegatorBondMore(
     collatorWalletAddress,
